// frontend-nx/apps/edu-hub/queries/__generated__/LoadAchievementOptionEvaluationScript.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: LoadAchievementOptionEvaluationScript
// ====================================================

export interface LoadAchievementOptionEvaluationScript_loadAchievementOptionEvaluationScript {
  __typename: "loadFileOutput";
  link: string;
}

export interface LoadAchievementOptionEvaluationScript {
  loadAchievementOptionEvaluationScript: LoadAchievementOptionEvaluationScript_loadAchievementOptionEvaluationScript | null;
}

export interface LoadAchievementOptionEvaluationScriptVariables {
  path: string;
}