// frontend-nx/apps/edu-hub/queries/__generated__/LoadAchievementOptionDocumentationTemplate.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: LoadAchievementOptionDocumentationTemplate
// ====================================================

export interface LoadAchievementOptionDocumentationTemplate_loadAchievementOptionDocumentationTemplate {
  __typename: "loadFileOutput";
  link: string;
}

export interface LoadAchievementOptionDocumentationTemplate {
  loadAchievementOptionDocumentationTemplate: LoadAchievementOptionDocumentationTemplate_loadAchievementOptionDocumentationTemplate | null;
}

export interface LoadAchievementOptionDocumentationTemplateVariables {
  path: string;
}